refactor: unify time constants
