--- packages/utils/src/timeConstants.ts
@@ -1,9 +1,0 @@
-/**
- * Milliseconds in a second
- */
-export const mts = 1000;
-
-/**
- * Milliseconds in a day
- */
-export const dayInMs = 86400000;
